+++ lib/rimless/karafka/base64_interchanger.rb
@@ -7,14 +7,14 @@ module Rimless
     #
     # rubocop:disable Security/MarshalLoad because we encode/decode the
     #   messages in our own controlled context
-    class Base64Interchanger
+    class Base64Interchanger < ::Karafka::Interchanger
       # Encode a binary Apache Kafka message(s) so they can be passed to the
       # Sidekiq +Rimless::ConsumerJob+.
       #
       # @param params_batch [Mixed] the raw message(s) to encode
       # @return [String] the marshaled+base64 encoded data
-      def self.encode(params_batch)
-        Base64.encode64(Marshal.dump(params_batch.to_a))
+      def encode(params_batch)
+        Base64.encode64(Marshal.dump(super))
       end
 
       # Decode the binary Apache Kafka message(s) so they can be processed by
@@ -22,8 +22,8 @@ module Rimless
       #
       # @param params_string [String] the marshaled+base64 encoded data
       # @return [Mixed] the unmarshaled+base64 decoded data
-      def self.decode(params_string)
-        Marshal.load(Base64.decode64(params_string))
+      def decode(params_string)
+        Marshal.load(Base64.decode64(super)).map(&:stringify_keys)
       end
     end
     # rubocop:enable Security/MarshalLoad